CA-266874: Fix infinite epoll wakeup loop
When a new connection is added to the QMP event thread, it writes a byte
to a pipe to wake up the event thread to add the new fd, but this byte
is never consumed so epoll_wait() will subsequently always return
immediately because the pipe is ready to read. Furthermore, when the
pipe is added to the monitor, wakeup() is called which immediately
triggers this loop. This causes xenopsd-xc to consume 100% of a CPU.

Fix this by removing the pipe entirely since it is not needed according
to the epoll_wait man page:
  "While one thread is blocked in a call to epoll_pwait(), it is
  possible for another thread to add a file descriptor to the
  waited-upon epoll instance. If the new file descriptor becomes ready,
  it will cause the epoll_wait() call to unblock.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>